// frontend/core-ui/types/Organisation.ts
export type Status = 'rejected_by_origin' | 'rejected_by_relation' | 'approved'
export type OrganisationRole = 'participating' | 'funding' | 'hosting'
export type OrganisationSource = 'RSD' | 'ROR' | 'MANUAL'

// organisation colums used in editOrganisation.createOrganisation
// NOTE! update when type CoreOrganisationProps changes
export const columsForCreate = [
  'parent', 'slug', 'primary_maintainer', 'name', 'ror_id', 'is_tenant', 'website', 'logo_id',
]
export type CoreOrganisationProps = {
  parent: string | null
  slug: string | null
  primary_maintainer?: string | null
  name: string
  ror_id: string | null
  is_tenant: boolean
  website: string | null
  logo_id: string | null
}

// organisation colums used in editOrganisation.updateOrganisation
// NOTE! update when type Organisation changes
export const columsForUpdate = [
  'id',
  'description',
  ...columsForCreate
]
export type Organisation = CoreOrganisationProps & {
  id: string | null
  // about page content created by maintainer
  description: string | null
  parent_names?: string
  rsd_path?: string
}

// adding source
export type SearchOrganisation = {
  id: string | null
  slug: string | null
  name: string,
  ror_id: string | null
  website: string | null,
  is_tenant: boolean
  logo_id: string | null
  primary_maintainer?: string | null
  parent: string | null
  parent_names?: string
  rsd_path?: string
  source: OrganisationSource
}

// extending with other props for software edit page
export type EditOrganisation = SearchOrganisation & {
  role?: OrganisationRole,
  position: number | null
  // new image to upload
  logo_b64: string | null
  logo_mime_type: string | null
  // source?: OrganisationSource
  status?: Status
  // only maintainers can edit values
  canEdit?: boolean
  // used for children to have complete rsd_path
  rsd_path?: string
}
// ...
export type OrganisationForOverview = Organisation & {
  id: string
  slug: string
  logo_id: string | null
  software_cnt: number | null
  project_cnt: number | null
  children_cnt: number | null
  release_cnt: number | null
  rsd_path: string
}